+++ requirements.txt
@@ -14,4 +14,3 @@ prometheus_client==0.7.1
 uWSGI==2.0.18
 py-healthcheck==1.9.0
 bandit==1.5.1
-
