+++ tests/unit/test_routes/test_api_key.py
@@ -2,18 +2,6 @@ from app.api.auth import blacklist_key
 from .helpers import get_api_key
 
 
-def test_apikey_commit_error(
-        module_client, module_db, fake_auth_from_oc, fake_commit_error):
-    client = module_client
-
-    response = client.post('api/v1/apikey', json=dict(
-        email="test@example.org",
-        password="supersecurepassword"
-    ))
-
-    assert (response.status_code == 500)
-
-
 def test_get_api_key(module_client, module_db, fake_auth_from_oc):
     client = module_client
 
@@ -25,6 +13,31 @@ def test_get_api_key(module_client, module_db, fake_auth_from_oc):
     assert (response.status_code == 200)
     assert (response.json['data'].get('email') == "test@example.org")
     assert (isinstance(response.json['data'].get('apikey'), str))
+
+
+def test_rotate_api_key(module_client, module_db, fake_auth_from_oc):
+    client = module_client
+
+    apikey = get_api_key(client)
+    response = client.post('api/v1/apikey/rotate', headers={'x-apikey': apikey})
+
+    assert (response.status_code == 200)
+    assert (isinstance(response.json['data'].get('email'), str))
+    assert (isinstance(response.json['data'].get('apikey'), str))
+    assert (response.json['data'].get('apikey') != apikey)
+
+
+def test_apikey_commit_error(
+        module_client, module_db, fake_auth_from_oc, fake_commit_error):
+    client = module_client
+
+    response = client.post('api/v1/apikey', json=dict(
+        email="test@example.com",
+        password="password"
+    ))
+    print(response.json)
+
+    assert (response.status_code == 500)
 
 
 def test_get_api_key_bad_password(module_client, module_db, fake_invalid_auth_from_oc):
@@ -66,18 +79,6 @@ def test_rotate_api_key_unauthorized(module_client, module_db):
     response = client.post('api/v1/apikey/rotate')
 
     assert (response.status_code == 401)
-
-
-def test_rotate_api_key(module_client, module_db, fake_auth_from_oc):
-    client = module_client
-
-    apikey = get_api_key(client)
-    response = client.post('api/v1/apikey/rotate', headers={'x-apikey': apikey})
-
-    assert (response.status_code == 200)
-    assert (isinstance(response.json['data'].get('email'), str))
-    assert (isinstance(response.json['data'].get('apikey'), str))
-    assert (response.json['data'].get('apikey') != apikey)
 
 
 def test_key_query_error(
